version 38.0
In version 38.0 I made the following changes to resolve the specific issues you mentioned:
Actimel image: Updated to the new link you provided.
Version number: Added a small caption in the top left (v38.0) to identify the version in print.
Fix mobile printing (main challenge):
I changed the printing approach. Instead of relying on zoom (which behaves differently across browsers), I set the body width to the physical width of an A4 page (297mm) and the height to 210mm.
I used flexbox in print to center the table exactly in the middle of the page.
I added a small inner padding to prevent the first or last column from being cut off.
The table is now set to fill 98% of the page width, which should make the most of the space and be clear.